reorganization of the file
