+++ assets/styles/display-selected-component/display_single_choice.scss
