+++ templates/modals/_modal_section.html.twig
@@ -1,7 +1,7 @@
 <div class="modal-section-container">	
 	<div class="modal-section-content">
 		<form method="post">
-			<a href="#" class="close-modal-section modal-section-trigger">&times;</a>
+			<a class="close-modal-section modal-section-trigger">&times;</a>
 			<h1>Create a new section</h1>
 			<div class="input-modal-section modal-section-trigger">
 				<input type="hidden" name="sectionName" value="section">
